requirements.txt
# Web Framework
Flask==3.0.0
Flask-Login==0.6.3
Werkzeug==3.0.1

# Database
pymongo==4.6.1
dnspython==2.4.2

# NLP & ML
nltk==3.8.1
spacy==3.7.2
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4

# Resume Parsing
PyMuPDF==1.23.8
python-docx==1.1.0

# AI
anthropic==0.18.1

# MLOps
mlflow==2.9.2

# Utilities
python-dotenv==1.0.0
Pillow==10.1.0
gunicorn==21.2.0
